🔧 chore: Update EC2 Deployment Workflow
Update the deployment workflow for EC2 in the `cd.yml` file. Changed the runner from macOS to Linux, and specified the use of an environment file for Docker container configuration. This change allows the deployment process to utilize a `.env` file located at `/volume/.env` on the EC2 instance.

Related issue: #30