♻️ refactor: Add role_id to Select in GetUserClasses
enhanced the functionality of the GetUserClasses function by adding class_users.role_id to the list of fields selected in the query. This improvement aims to provide a more detailed and nuanced view of the classes a user is associated with, by including the role they hold within each class.

Related issue: #124