♻️ refactor: Exclude createdAt and pid from User Info in Frontend Response
In an effort to streamline the data transmitted to the frontend and enhance data privacy, this commit implements a crucial change in how user information is handled and sent. Specifically, we have modified the user data response to exclude the createdAt and pid (personal identification) fields when sending login user information to the frontend.

Related issue: #171