✨ feat: Assign Admin Role to User on Class Creation
Implemented logic in the ClassService to assign the admin role (role_id: 2) to the user who creates a new class. This update involves saving the class creator as an admin in the class_user table. The Save method in ClassUserRepository has been utilized to achieve this. This change ensures proper role management in class creation, enhancing the access control within our application.

Related issue: #166