♻️ refactor: Remove unused database-related files
This commit significantly tidies up the project repository by removing files related to the database that are currently not in use. Efforts to streamline our codebase led to the identification of several legacy files, configurations, and scripts which are either deprecated or have been replaced by more efficient solutions. The decision to remove these files was made to enhance the maintainability of the project and reduce potential confusion among new developers regarding the project's database architecture.

Related issue: #247