🔥remove: Delete role.go, an unused file
Removed the file role.go, which was no longer used within the project. This file previously contained code related to defining role models, but it was deemed unnecessary and unused in the current state of the project. By removing this file, we aim to declutter the codebase and improve maintainability.

Related issue: #247